packages/curate-web/pyproject.toml
[project]
name = "curate-web"
version = "0.1.0"
description = "FastAPI editorial dashboard for Curate."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "curate-common",
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.41.0",
    "msal>=1.34.0",
    "jinja2>=3.1.6",
    "httpx>=0.28.1",
    "sse-starlette>=3.2.0",
    "itsdangerous>=2.2.0",
    "azure-servicebus>=7.14.0",
    "azure-monitor-opentelemetry>=1.8.6",
]

[project.scripts]
curate-web = "curate_web.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/curate_web"]
